Refactored project to new name: Cacao Added color palette Added fonts Added button resource Added logo Added activity_welcome.xml
