+++ app/src/androidTest/java/dev/prognitio/cacao/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package dev.prognitio.congressionalapptempname;
+package dev.prognitio.cacao;
 
 import android.content.Context;
 
@@ -21,6 +21,6 @@ public class ExampleInstrumentedTest {
     public void useAppContext() {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
-        assertEquals("dev.prognitio.congressionalapptempname", appContext.getPackageName());
+        assertEquals("dev.prognitio.cacao", appContext.getPackageName());
     }
 }
