+++ app/src/main/java/dev/prognitio/cacao/WelcomeActivity.java
@@ -1,15 +1,14 @@
-package dev.prognitio.congressionalapptempname;
+package dev.prognitio.cacao;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
-import android.view.Window;
 
-public class MainActivity extends AppCompatActivity {
+public class WelcomeActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.activity_welcome);
     }
 }
