+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">CongressionalAppTempName</string>
+    <string name="app_name">Cacao</string>
 </resources>
