+++ app/src/test/java/dev/prognitio/cacao/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package dev.prognitio.congressionalapptempname;
+package dev.prognitio.cacao;
 
 import org.junit.Test;
