+++ settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "CongressionalAppTempName"
+rootProject.name = "Cacao"
 include ':app'
